Allow for a time adjustment to timestamps
Test systems, at least robot framework, provide a timestamp without
timezone information, so timestamps are usually in local time and may
or may not include daylight savings.
When saving to database this is best in utc time. To allow the user decide
what they want to save in, provide an offset which they can pass in from
the command line. Have the offset in seconds for flexibility.