Use the TimeAdjust class to adjust timestamps to fixed and/or system offset
In the archiver change all usages of previously fixed _time_adjust_secs
to use the TimeAdjust class which dynamically calculates the adjustment
offset.
In the listener we can delegate the calculation back to the archiver now
that it is also capable of adjusting to system timezone and/or daylight
savings.