Add metadata to suite test run if timestamps are adjusted
If either or both of
--time-adjust-with-system-timezone
--time-adjust-secs
are set (i.e. included on command line, or set in a listener) then add a metadata as a trace.

e.g with command line
output_parser.py --time-adjust-secs -3600 --time-adjust-with-system-timezone ...
the following values would be added to suite_metadata table
time_adjust_secs with value -3600
time_adjust_secs_total with -10800.
This example is mimicking adding daylight savings (1hr = 3600 secs) onto
a system offset secs of 7200 (GMT+2).

If the time-adjust-secs is not set or zero, no entry is made to the table for
time_adjust_secs.
If the time-adjust-with-system-timezone is false (option not included) no
entry is made to the table for time_adjust_secs_total.