Allow  arbitrary table-field combinations against test suite and run
To provide a way for users of TestArchiver to extend the data, add a
special metadata name table#<table-name>:{json-fields} where json-fields
is of format {"field1":"value1", "field2":"value2"} and has to
correspond exactly to the fields in the table-name table or naturally
an exception will be thrown.
The creation of the table is left to the user, as TestArchiver can not
know what the table is.